Trakk ut vilkaarOgBegrunnelse i egen fil
