+++ src/sider/ftrl/saksbehandling/initialStegs.tsx
@@ -1,7 +1,7 @@
 import { FANE_STATUS } from "../../../felleskomponenter/stegvelger";
 import { VurderingStart } from "./stegKomponenter/vurderingStart";
 import { VurderingVirksomhet } from "./stegKomponenter/vurderingVirksomhet";
-import { VurderingBestemmelse } from "./stegKomponenter/vurderingBestemmelse";
+import { VurderingBestemmelse } from "./stegKomponenter/vurderingBestemmelse/vurderingBestemmelse";
 import { VurderingPerioder } from "./stegKomponenter/vurderingPerioder";
 import { VurderingTrygdeavgift } from "./stegKomponenter/vurderingTrygdeavgift";
 import { VurderingVedtak } from "./stegKomponenter/vurderingVedtak";
