+++ src/sider/ftrl/saksbehandling/stegKomponenter/vurderingBestemmelse/komponenter/flytFinnesIkke.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import * as Nav from "../../../navFrontend";
+import * as Nav from "../../../../../../navFrontend";
 
 export const FlytFinnesIkke = () => (
   <Nav.AlertStripeAdvarsel>
